todoitem/backend/requirements.txt
flask
flask-pymongo
flask-cors
python-dotenv